Move HttpRequest outside node directory
